requirements.txt
modal==1.4.2
protobuf>=6,<7
fastapi==0.110.0
python-multipart==0.0.27
anthropic==0.39.0
elevenlabs==1.7.0
twilio==9.0.0
python-dotenv==1.0.1
pytest==8.0.0
pytest-asyncio==0.23.0
httpx==0.27.0
tzdata==2026.2
google-api-python-client==2.196.0
google-auth==2.52.0
google-auth-oauthlib==1.4.0
